Add more compile time and runtime checks
Task: #24

+++ benchmarks/complexity_bench.cpp
@@ -22,7 +22,7 @@ void num_of_tasks(benchmark::State &state)
         (void) _;
         benchmark::DoNotOptimize(runners::block_on([&state, &task_coro]() -> task<std::size_t> {
             std::vector<::coschedula::task<std::size_t>> tasks;
-            tasks.reserve(state.range(0));
+            tasks.reserve(static_cast<std::size_t>(state.range(0)));
             for (std::int64_t i = 0; i < state.range(0); ++i) {
                 tasks.push_back(task_coro());
             }
@@ -52,7 +52,7 @@ void num_of_tasks_async(benchmark::State &state)
         (void) _;
         benchmark::DoNotOptimize(runners::block_on([&state, &task_coro]() -> task<std::size_t> {
             std::vector<::coschedula::task<std::size_t>> tasks;
-            tasks.reserve(state.range(0));
+            tasks.reserve(static_cast<std::size_t>(state.range(0)));
             for (std::int64_t i = 0; i < state.range(0); ++i) {
                 tasks.push_back(task_coro());
             }
